refactor(observability): remove unused NodeLogStreamTracker
